Wszystko dziala, ale bedzie refactoring probably

+++ knn/k-NN-Algorithm/k-NN-Algorithm/MinkowskiDistance.cpp
@@ -5,7 +5,7 @@ MinkowskiDistance::MinkowskiDistance() {
 	this->p = 3;
 }
 
-MinkowskiDistance::MinkowskiDistance(unsigned int p) {
+MinkowskiDistance::MinkowskiDistance(double p) {
 	this->p = p;
 }
